# driving/api-rest/src/main/java/com/mercadona/alejandro/dev/driving/controllers/adapters/api/ProductsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mercadona.alejandro.dev.driving.controllers.adapters.api;

import com.mercadona.alejandro.dev.driving.controllers.adapters.model.ErrorResource;
import com.mercadona.alejandro.dev.driving.controllers.adapters.model.ProductPageResponse;
import com.mercadona.alejandro.dev.driving.controllers.adapters.model.ProductRequest;
import com.mercadona.alejandro.dev.driving.controllers.adapters.model.ProductResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.lang.Nullable;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-14T05:55:16.740231+01:00[Europe/Madrid]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "products", description = "Product Controller")
public interface ProductsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_PRODUCT = "/products";
    /**
     * POST /products : Method to add a new product
     * This method create a new product
     *
     * @param productRequest productRequest (optional)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not Found (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Method to add a new product",
        description = "This method create a new product",
        tags = { "products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            })
        },
        security = {
            @SecurityRequirement(name = "adfs")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ProductsApi.PATH_CREATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ProductResponse> createProduct(
        @Parameter(name = "ProductRequest", description = "productRequest") @Valid @RequestBody(required = false) @Nullable ProductRequest productRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"name\" : \"name\", \"id\" : 0, \"state\" : \"ACTIVE\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_CODE_BAR_PRODUCT = "/products/product-bar-code/{codeBar}";
    /**
     * GET /products/product-bar-code/{codeBar} : Method to retrieve one product
     * This method retrieves a product specific by bar code
     *
     * @param codeBar codeBar (required)
     * @param locale Language locale (optional, default to es-ES)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not Found (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getCodeBarProduct",
        summary = "Method to retrieve one product",
        description = "This method retrieves a product specific by bar code",
        tags = { "products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            })
        },
        security = {
            @SecurityRequirement(name = "adfs")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_GET_CODE_BAR_PRODUCT,
        produces = { "application/json" }
    )
    default ResponseEntity<ProductResponse> getCodeBarProduct(
        @NotNull @Parameter(name = "codeBar", description = "codeBar", required = true, in = ParameterIn.PATH) @PathVariable("codeBar") String codeBar,
        @Pattern(regexp = "^[a-z]{2}([_-][A-Z]{2})?$") @Parameter(name = "locale", description = "Language locale", in = ParameterIn.QUERY) @Valid @RequestParam(value = "locale", required = false, defaultValue = "es-ES") String locale
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"name\" : \"name\", \"id\" : 0, \"state\" : \"ACTIVE\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_PRODUCT_BY_ID = "/products/{productId}";
    /**
     * GET /products/{productId} : Method to retrieve one product
     * This method retrieves a product specific by identifier
     *
     * @param productId Product Id (required)
     * @param locale Language locale (optional, default to es-ES)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not Found (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Method to retrieve one product",
        description = "This method retrieves a product specific by identifier",
        tags = { "products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            })
        },
        security = {
            @SecurityRequirement(name = "adfs")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_GET_PRODUCT_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<ProductResponse> getProductById(
        @NotNull @Parameter(name = "productId", description = "Product Id", required = true, in = ParameterIn.PATH) @PathVariable("productId") Long productId,
        @Pattern(regexp = "^[a-z]{2}([_-][A-Z]{2})?$") @Parameter(name = "locale", description = "Language locale", in = ParameterIn.QUERY) @Valid @RequestParam(value = "locale", required = false, defaultValue = "es-ES") String locale
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"name\" : \"name\", \"id\" : 0, \"state\" : \"ACTIVE\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_PRODUCT_COLLECTION = "/products";
    /**
     * GET /products : Method to retrieve a collection of products
     * This method retrieves a collection of products
     *
     * @param firstPage Requested page (first page is page 1). If not specified 1 is used. [Cannot be negative or zero] (optional, default to 1)
     * @param pageSize Number of retrieved elements per page. if not specified 10 is used. [Cannot be less than 1] (optional, default to 10)
     * @param sort Sort of the query. [sort&#x3D;+id] will order asc by id, or [sort&#x3D;+id] desc by id). Default unsorted. (optional)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not Found (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getProductCollection",
        summary = "Method to retrieve a collection of products",
        description = "This method retrieves a collection of products",
        tags = { "products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductPageResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "403", description = "Forbidden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResource.class))
            })
        },
        security = {
            @SecurityRequirement(name = "adfs")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_GET_PRODUCT_COLLECTION,
        produces = { "application/json" }
    )
    default ResponseEntity<ProductPageResponse> getProductCollection(
        @Parameter(name = "firstPage", description = "Requested page (first page is page 1). If not specified 1 is used. [Cannot be negative or zero]", in = ParameterIn.QUERY) @Valid @RequestParam(value = "firstPage", required = false, defaultValue = "1") Integer firstPage,
        @Parameter(name = "pageSize", description = "Number of retrieved elements per page. if not specified 10 is used. [Cannot be less than 1]", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize,
        @Parameter(name = "sort", description = "Sort of the query. [sort=+id] will order asc by id, or [sort=+id] desc by id). Default unsorted.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) @Nullable String sort
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"pagination\" : { \"requestedPage\" : 6, \"requestedSize\" : 1, \"totalResults\" : 5, \"previousPage\" : \"previousPage\", \"retrievedResults\" : 5, \"nextPage\" : \"nextPage\" }, \"data\" : [ { \"code\" : \"code\", \"name\" : \"name\", \"id\" : 0, \"state\" : \"ACTIVE\" }, { \"code\" : \"code\", \"name\" : \"name\", \"id\" : 0, \"state\" : \"ACTIVE\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"description\" : \"description\", \"details\" : [ \"details\", \"details\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}